Update Jaeger propagator. (#663)
