+++ propagator/jaeger/test/jaeger_test.rb
@@ -7,20 +7,11 @@
 require 'test_helper'
 
 describe OpenTelemetry::Propagator::Jaeger do
-  describe '#text_map_extractor' do
-    it 'returns an instance of TextMapExtractor' do
-      extractor = OpenTelemetry::Propagator::Jaeger.text_map_extractor
-      _(extractor).must_be_instance_of(
-        OpenTelemetry::Propagator::Jaeger::TextMapExtractor
-      )
-    end
-  end
-
-  describe '#text_map_injector' do
-    it 'returns an instance of TextMapInjector' do
-      injector = OpenTelemetry::Propagator::Jaeger.text_map_injector
-      _(injector).must_be_instance_of(
-        OpenTelemetry::Propagator::Jaeger::TextMapInjector
+  describe '#text_map_propagator' do
+    it 'returns an instance of TextMapPropagator' do
+      propagator = OpenTelemetry::Propagator::Jaeger.text_map_propagator
+      _(propagator).must_be_instance_of(
+        OpenTelemetry::Propagator::Jaeger::TextMapPropagator
       )
     end
   end
